add & use scrub datasource workflow
