+++ front/pages/api/poke/workspaces/[wId]/data_sources/[name]/index.ts
@@ -1,4 +1,3 @@
-import { Storage } from "@google-cloud/storage";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { getSession, getUserFromSession } from "@app/lib/auth";
@@ -7,8 +6,7 @@ import { CoreAPI } from "@app/lib/core_api";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import { DataSource, Workspace } from "@app/lib/models";
 import { apiError, withLogging } from "@app/logger/withlogging";
-
-const { DUST_DATA_SOURCES_BUCKET, SERVICE_ACCOUNT } = process.env;
+import { launchScrubDataSourceWorkflow } from "@app/poke/temporal/client";
 
 export type DeleteDataSourceResponseBody = {
   success: true;
@@ -43,25 +41,6 @@ async function handler(
 
   switch (req.method) {
     case "DELETE":
-      if (!SERVICE_ACCOUNT) {
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: "Could not find the service account for GCP config.",
-          },
-        });
-      }
-      if (!DUST_DATA_SOURCES_BUCKET) {
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: "Could not find the datasource bucket config.",
-          },
-        });
-      }
-
       const { wId } = req.query;
       if (!wId || typeof wId !== "string") {
         return apiError(req, res, {
@@ -134,31 +113,10 @@ async function handler(
 
       await dataSource.destroy();
 
-      const storage = new Storage({ keyFilename: SERVICE_ACCOUNT });
-
-      const [files] = await storage
-        .bucket(DUST_DATA_SOURCES_BUCKET)
-        .getFiles({ prefix: dustAPIProjectId });
-
-      const chunkSize = 32;
-      const chunks = [];
-      for (let i = 0; i < files.length; i += chunkSize) {
-        chunks.push(files.slice(i, i + chunkSize));
-      }
-
-      for (let i = 0; i < chunks.length; i++) {
-        const chunk = chunks[i];
-        if (!chunk) {
-          continue;
-        }
-        await Promise.all(
-          chunk.map((f) => {
-            return (async () => {
-              await f.delete();
-            })();
-          })
-        );
-      }
+      await launchScrubDataSourceWorkflow({
+        workspaceId: workspace.sId,
+        dustAPIProjectId,
+      });
 
       return res.status(200).json({ success: true });
