+++ front/poke/temporal/client.ts
@@ -1,0 +1,23 @@
+import { getTemporalClient } from "@app/lib/temporal";
+
+import { scrubDataSourceWorkflow } from "./workflows";
+
+export async function launchScrubDataSourceWorkflow({
+  workspaceId,
+  dustAPIProjectId,
+}: {
+  workspaceId: string;
+  dustAPIProjectId: string;
+}) {
+  const client = await getTemporalClient();
+
+  await client.workflow.start(scrubDataSourceWorkflow, {
+    args: [
+      {
+        dustAPIProjectId,
+      },
+    ],
+    taskQueue: "poke-queue",
+    workflowId: `scrub-data-source-${workspaceId}-${dustAPIProjectId}`,
+  });
+}
